Feat: time complexity of post data processing time changed from n^2 to n.
