+++ app/_components/postPreview/AnimalPostPreview.tsx
@@ -32,22 +32,22 @@ export default function AnimalPostPreview() {
   const [selectedAnimal, setSelectedAnimal] = useState('')
   const [posts, setPosts] = useState<Array<postData>>()
 
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const res = await api.get(
-          `/posts?${selectedAnimal !== '' ? 'animal=' + selectedAnimal + '&' : ''}page=1&pageSize=3`,
-        )
-        const posts = await res.data
-        console.log(posts, 'animal')
+  // useEffect(() => {
+  //   const fetchData = async () => {
+  //     try {
+  //       const res = await api.get(
+  //         `/posts?${selectedAnimal !== '' ? 'animal=' + selectedAnimal + '&' : ''}page=1&pageSize=3`,
+  //       )
+  //       const posts = await res.data
+  //       console.log(posts, 'animal')
 
-        setPosts(posts)
-      } catch (err) {
-        console.error(err)
-      }
-    }
-    fetchData()
-  }, [selectedAnimal])
+  //       setPosts(posts)
+  //     } catch (err) {
+  //       console.error(err)
+  //     }
+  //   }
+  //   fetchData()
+  // }, [selectedAnimal])
   const handleAnimalSelect = (animal: string) => {
     setSelectedAnimal(() => animal)
   }
@@ -63,7 +63,7 @@ export default function AnimalPostPreview() {
           <Next />
         </Link>
       </div>
-      <div className={style.animalContainer}>
+      {/* <div className={style.animalContainer}>
         <Button
           text={'전체'}
           mainColor={selectedAnimal === '' ? 'green' : 'grey'}
@@ -104,16 +104,15 @@ export default function AnimalPostPreview() {
             handleAnimalSelect('폭스')
           }}
         />
-      </div>
+      </div> */}
       <div className={style.animalPostContainer}>
-        {posts?.map((postData, index: number) => {
+        {/* {posts?.map((postData: postData, index: number) => {
           return (
             <Post
               key={index}
               profile={{
-                image: null,
-                // name: postData.author.credential.nickname,
-                name: 'nickname',
+                image: postData.author?.profileItems,
+                name: postData.author?.credential.nickname, //TODO: 02.26 회의 후 탈퇴한 회원 정보 처리
                 animal: postData.animalOfAuthor,
                 isAnonymous: postData.isAnonymous,
               }}
@@ -130,7 +129,7 @@ export default function AnimalPostPreview() {
               }}
             />
           )
-        })}
+        })} */}
       </div>
     </section>
   )
