+++ app/_components/postPreview/PostPreview.tsx
@@ -4,7 +4,10 @@ import Link from '@/node_modules/next/link'
 
 import Next from '@/public/assets/next.svg'
 import { ReactNode } from 'react'
-import Posts from './_components/Posts'
+
+import { postData } from './_types/responseType'
+import api from '@/app/_api/commonApi'
+import Post from './_components/Post'
 
 type Props = {
   title: string
@@ -12,7 +15,23 @@ type Props = {
   icon?: ReactNode
 }
 
-export default function PostPreview({ title, href, icon }: Props) {
+export default async function PostPreview({ title, href, icon }: Props) {
+  const fetchData = async () => {
+    try {
+      const { data } = await api.get(
+        // `/posts?page=1&pageSize=2${title === '연애 TIP' ? '&category=연애' : ''}`,
+        `/posts?page=1&pageSize=2`,
+      )
+
+      console.log(data, title, 'post data')
+
+      return data
+    } catch (err) {
+      console.error(err, 'post error')
+    }
+  }
+  const postData = await fetchData()
+
   return (
     <section className={style.container}>
       <div className={style.titleContainer}>
@@ -25,8 +44,30 @@ export default function PostPreview({ title, href, icon }: Props) {
         </Link>
       </div>
       <div className={style.postContainer}>
-        {/* @ts-expect-error Async Server Component */}
-        <Posts title={title} />
+        {postData?.map((postData: postData, index: number) => {
+          return (
+            <Post
+              key={index}
+              profile={{
+                image: postData.author?.profileItems,
+                name: postData.author?.credential.nickname,
+                animal: postData.animalOfAuthor,
+                isAnonymous: postData.isAnonymous,
+              }}
+              post={{
+                id: postData.id,
+                category: postData.category,
+                tag: postData.tag,
+                time: postData.createdAt,
+                title: postData.title,
+                content: postData.content,
+                likes: postData.likes,
+                comments: postData.commentNum,
+                views: postData.views,
+              }}
+            />
+          )
+        })}
       </div>
     </section>
   )
