+++ app/_components/postPreview/postPreview.module.scss
@@ -13,15 +13,8 @@
 }
 
 .title {
-  display: flex;
-  align-items: center;
-  height: 100%;
-  color: #000;
-  font-family: Pretendard;
   font-size: 19px;
-  font-style: normal;
   font-weight: 700;
-  line-height: normal;
 }
 .postContainer {
   display: flex;
